PSP-1254 Apply MOTI DB Schema Standards

+++ backend/api/Areas/Admin/Models/Parcel/BuildingEvaluationModel.cs
@@ -5,7 +5,7 @@ namespace Pims.Api.Areas.Admin.Models.Parcel
     public class BuildingEvaluationModel : Api.Models.BaseModel
     {
         #region Properties
-        public int BuildingId { get; set; }
+        public long BuildingId { get; set; }
 
         public DateTime Date { get; set; }
