+++ backend/api/Areas/Admin/Models/Parcel/ParcelFiscalModel.cs
@@ -3,9 +3,9 @@ namespace Pims.Api.Areas.Admin.Models.Parcel
     public class ParcelFiscalModel : Api.Models.BaseModel
     {
         #region Properties
-        public int ParcelId { get; set; }
+        public long ParcelId { get; set; }
 
-        public int FiscalYear { get; set; }
+        public long FiscalYear { get; set; }
 
         public string Key { get; set; }
