+++ backend/api/Areas/Admin/Models/Parcel/PartialPropertyModel.cs
@@ -5,11 +5,11 @@ namespace Pims.Api.Areas.Admin.Models.Parcel
     public class PartialPropertyModel : Model.BaseModel
     {
         #region Properties
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int StatusId { get; set; }
+        public long StatusId { get; set; }
 
-        public int ClassificationId { get; set; }
+        public long ClassificationId { get; set; }
 
         public double Latitude { get; set; }
