+++ backend/api/Areas/Project/Models/Dispose/BuildingModel.cs
@@ -7,19 +7,19 @@ namespace Pims.Api.Areas.Project.Models.Dispose
     public class BuildingModel : PropertyModel
     {
         #region Properties
-        public int ParcelId { get; set; }
+        public long ParcelId { get; set; }
 
-        public int BuildingConstructionTypeId { get; set; }
+        public long BuildingConstructionTypeId { get; set; }
 
         public string BuildingConstructionType { get; set; }
 
         public int BuildingFloorCount { get; set; }
 
-        public int BuildingPredominateUseId { get; set; }
+        public long BuildingPredominateUseId { get; set; }
 
         public string BuildingPredominateUse { get; set; }
 
-        public int BuildingOccupantTypeId { get; set; }
+        public long BuildingOccupantTypeId { get; set; }
 
         public string BuildingOccupantType { get; set; }
