+++ backend/api/Areas/Project/Models/Dispose/ProjectModel.cs
@@ -12,7 +12,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The primary key to identify the project.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
         /// <summary>
         /// get/set - A unique project number to identify the project.
         /// </summary>
@@ -40,7 +40,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The foreign key to the workflow.
         /// </summary>
-        public int WorkflowId { get; set; }
+        public long WorkflowId { get; set; }
 
         /// <summary>
         /// get/set - The code of the workflow.
@@ -50,7 +50,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The foreign key to the project status.
         /// </summary>
-        public int StatusId { get; set; }
+        public long StatusId { get; set; }
 
         /// <summary>
         /// get/set - The code of the project status.
@@ -65,7 +65,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The foreign key to the project risk.
         /// </summary>
-        public int RiskId { get; set; }
+        public long RiskId { get; set; }
 
         /// <summary>
         /// get/set - The project risk name.
@@ -75,7 +75,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The foreign key to the tier level.
         /// </summary>
-        public int TierLevelId { get; set; }
+        public long TierLevelId { get; set; }
 
         /// <summary>
         /// get/set - The tier level of the project.
@@ -130,7 +130,7 @@ namespace Pims.Api.Areas.Project.Models.Dispose
         /// <summary>
         /// get/set - The foreign key to the owning agency.
         /// </summary>
-        public int AgencyId { get; set; }
+        public long AgencyId { get; set; }
 
         /// <summary>
         /// get/set - The owning agency name.
