+++ backend/api/Areas/Project/Models/Search/ProjectPropertyModel.cs
@@ -8,12 +8,12 @@ namespace Pims.Api.Areas.Project.Models.Search
         /// <summary>
         /// get/set - The primary key to identify the property.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// get/set - The foreign key to the property type [Land, Building].
         /// </summary>
-        public int PropertyTypeId { get; set; }
+        public long PropertyTypeId { get; set; }
 
         /// <summary>
         /// get/set - property name.
@@ -28,7 +28,7 @@ namespace Pims.Api.Areas.Project.Models.Search
         /// <summary>
         /// get/set - The foreign key to the owning agency.
         /// </summary>
-        public int AgencyId { get; set; }
+        public long AgencyId { get; set; }
 
         /// <summary>
         /// get/set - The owning agency name.
