+++ backend/api/Areas/Property/Models/Parcel/ParcelEvaluationModel.cs
@@ -6,7 +6,7 @@ namespace Pims.Api.Areas.Property.Models.Parcel
     public class ParcelEvaluationModel : Model.BaseModel
     {
         #region Properties
-        public int ParcelId { get; set; }
+        public long ParcelId { get; set; }
 
         public DateTime Date { get; set; }
