+++ backend/api/Areas/Property/Models/Parcel/ParcelFiscalModel.cs
@@ -6,7 +6,7 @@ namespace Pims.Api.Areas.Property.Models.Parcel
     public class ParcelFiscalModel : Model.BaseModel
     {
         #region Properties
-        public int ParcelId { get; set; }
+        public long ParcelId { get; set; }
 
         public int FiscalYear { get; set; }
