+++ backend/api/Areas/Property/Models/Parcel/PartialPropertyModel.cs
@@ -5,13 +5,13 @@ namespace Pims.Api.Areas.Property.Models.Parcel
     public class PartialPropertyModel : Model.BaseModel
     {
         #region Properties
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
-        public int StatusId { get; set; }
+        public long StatusId { get; set; }
 
-        public int ClassificationId { get; set; }
+        public long ClassificationId { get; set; }
 
         public double Latitude { get; set; }
