+++ backend/api/Areas/Tools/Models/Import/BuildingEvaluationModel.cs
@@ -5,7 +5,7 @@ namespace Pims.Api.Areas.Tools.Models.Import
     public class BuildingEvaluationModel : Api.Models.BaseModel
     {
         #region Properties
-        public int BuildingId { get; set; }
+        public long BuildingId { get; set; }
 
         public DateTime Date { get; set; }
