+++ backend/api/Areas/Tools/Models/Import/ProjectModel.cs
@@ -9,7 +9,7 @@ namespace Pims.Api.Areas.Tools.Models.Import
         /// <summary>
         /// get/set - The primary key provides a unique identity for the project.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// get/set - A unique identity for the project.
@@ -44,7 +44,7 @@ namespace Pims.Api.Areas.Tools.Models.Import
         /// <summary>
         /// get/set - The foreign key to the owning agency.
         /// </summary>
-        public int AgencyId { get; set; }
+        public long AgencyId { get; set; }
 
         /// <summary>
         /// get/set - The owning agency name.
@@ -69,7 +69,7 @@ namespace Pims.Api.Areas.Tools.Models.Import
         /// <summary>
         /// get/set - The foreign key to the project status.
         /// </summary>
-        public int StatusId { get; set; }
+        public long StatusId { get; set; }
 
         /// <summary>
         /// get/set - The project status.
@@ -94,7 +94,7 @@ namespace Pims.Api.Areas.Tools.Models.Import
         /// <summary>
         /// get/set - The foreign key to the tier level.
         /// </summary>
-        public int TierLevelId { get; set; }
+        public long TierLevelId { get; set; }
 
         /// <summary>
         /// get/set - The tier level.
